Change latlong to doubles instead of floats for precision

+++ RestServer/database/migrations/2015_05_01_084155_create_stations_table.php
@@ -18,8 +18,8 @@ class CreateStationsTable extends Migration {
 			$table->integer('number')->unique;
 			$table->string('name');
 			$table->string('address');
-			$table->float('longitude');
-			$table->float('latitude');
+			$table->double('longitude');
+			$table->double('latitude');
 			$table->boolean('banking');
 			$table->boolean('bonus');
 			$table->string('status');
